Backend/requirements.txt
django==4.2.0
numpy==1.26.0
pillow==10.3.0
channels==4.1.0